+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## [0.19.1] - 2023-11-27
+
+### Added
+
+- `EafPlus.get_annotations` and `EafPlus.get_annotations_and_intervals` now have a new `drop_empty_tiers` argument that defaults to `True`. If `True`, empty tiers are dropped from the output. If `False`, empty tiers are included in the output as completely empty annotations.
+
+### Fixed
+
+- `EafPlus.get_flattened_annotations_for_tier` (prev.`EafPlus.get_full_annotations_for_tier`) now correctly handles tiers with empty annotations (i.e., the annotation value is an empty string) as empty and these tiers get an empty row in the output table.
+- `EafPlus.get_annotations_and_intervals` now correctly handles empty tiers and assigns them code_num of <NA> instead of -1 to differentiate them from filled-in annotations outside any code intervals.
+- `EafPlus.get_annotations_and_intervals` has a docstring now.
+
+### Changed
+
+- `EafPlus.get_full_annotations_for_tier` renamed to `EafPlus.get_flattened_annotations_for_tier` because the `full` part was unclear - full as compared to what? Also, `flattened` implies that there is a hierarchy involved, which implicitly says what the function does - returns a table where each row is an utterance and all its annotations.
+
 ## [0.19.0] - 2023-11-21
 
 ### Added